+++ src/green/InterfacesImpl.hpp
@@ -38,15 +38,13 @@
 #include <boost/numeric/odeint.hpp>
 
 #include "utils/MathUtils.hpp"
+#include "utils/RungeKutta4.hpp"
+
+/*! \file InterfacesImpl.hpp */
 
 namespace pcm {
 namespace green {
 namespace detail {
-
-/*! \typedef StateType
- *  \brief state vector for the differential equation integrator
- */
-typedef std::vector<double> StateType;
 
 /*! \typedef ProfileEvaluator
  *  \brief sort of a function pointer to the dielectric profile evaluation function
@@ -75,16 +73,18 @@ struct IntegratorParameters {
  *  Provides a handle to the system of differential equations for the integrator.
  *  The dielectric profile comes in as a boost::function object.
  */
-class LnTransformedRadial {
+class LnTransformedRadial __final : public pcm::utils::detail::ODESystem<2> {
+public:
+  /*! Type of the state vector of the ODE */
+  typedef pcm::utils::detail::ODESystem<2>::StateType StateType;
+  /*! Constructor from profile evaluator and angular momentum */
+  LnTransformedRadial(const ProfileEvaluator & e, int lval) : eval_(e), l_(lval) {}
+
 private:
   /*! Dielectric profile function and derivative evaluation */
   ProfileEvaluator eval_;
   /*! Angular momentum */
   int l_;
-
-public:
-  /*! Constructor from profile evaluator and angular momentum */
-  LnTransformedRadial(const ProfileEvaluator & e, int lval) : eval_(e), l_(lval) {}
   /*! Provides a functor for the evaluation of the system
    *  of first-order ODEs needed by Boost.Odeint
    *  The second-order ODE and the system of first-order ODEs
@@ -93,7 +93,7 @@ public:
    *  \param[out] drhodr state vector holding the first and second derivative
    *  \param[in] y logarithmic position on the integration grid
    */
-  void operator()(const StateType & rho, StateType & drhodr, const double y) {
+  virtual void RHS(const StateType & rho, StateType & drhodr, const double y) const {
     // Evaluate the dielectric profile
     double eps = 0.0, epsPrime = 0.0;
     pcm::tie(eps, epsPrime) = eval_(std::exp(y));
@@ -107,76 +107,49 @@ public:
 };
 } // namespace detail
 
-using detail::ProfileEvaluator;
 using detail::IntegratorParameters;
+using detail::ProfileEvaluator;
 
-/*! \file InterfacesImpl.hpp
- *  \class RadialFunction
+/*! \class RadialFunction
  *  \brief represents solutions to the radial 2nd order ODE
  *  \author Roberto Di Remigio
- *  \date 2015
- *  \tparam StateVariable type of the state variable used in the ODE solver
- *  \tparam ODESystem system of 1st order ODEs replacing the 2nd order ODE
- *  \tparam IndependentSolution encodes which type of radial solution
+ *  \date 2018
+ *  \tparam ODE system of 1st order ODEs replacing the 2nd order ODE
+ *
+ * The sign of the integration interval, i.e. the sign of the difference
+ * between the minimum and the maximum of the interval, is used to discriminate
+ * between the two independent solutions (zeta-type and omega-type) of the
+ * differential equation:
+ *   - When the sign is positive (y_min_ < y_max_), we are integrating **forwards**
+ * for the zeta-type solution.
+ *   - When the sign is negative (y_max_ < y_min_), we are integrating **backwards**
+ * for the omega-type solution.
+ *
+ * This is based on an idea Luca had in Wuhan/Chongqing for the planar
+ * diffuse interface.
+ * With respect to the previous, much more heavily template-d implementation,
+ * it introduces a bit more if-else if-else branching in the function_impl and
+ * derivative_impl functions.
+ * This is largely offset by the better readability and reduced code duplication.
  */
-template <typename StateVariable,
-          typename ODESystem,
-          template <typename, typename> class IndependentSolution>
-class RadialFunction __final {
+template <typename ODE> class RadialFunction __final {
 public:
-  RadialFunction() : solution_(IndependentSolution<StateVariable, ODESystem>()) {}
+  RadialFunction() : L_(0), y_min_(0.0), y_max_(0.0), y_sign_(1) {}
   RadialFunction(int l,
-                 double y0,
-                 double yinf,
+                 double ymin,
+                 double ymax,
                  const ProfileEvaluator & eval,
                  const IntegratorParameters & parms)
-      : solution_(IndependentSolution<StateVariable, ODESystem>(l,
-                                                                y0,
-                                                                yinf,
-                                                                eval,
-                                                                parms)) {}
-  ~RadialFunction() {}
-  /*! \brief Returns value of function and its first derivative at given point
-   *  \param[in] point evaluation point
-   */
+      : L_(l),
+        y_min_(ymin),
+        y_max_(ymax),
+        y_sign_(pcm::utils::sign(y_max_ - y_min_)) {
+    compute(eval, parms);
+  }
   pcm::tuple<double, double> operator()(double point) const {
-    return solution_(point);
+    return pcm::make_tuple(function_impl(point), derivative_impl(point));
   }
   friend std::ostream & operator<<(std::ostream & os, RadialFunction & obj) {
-    os << obj.solution_;
-    return os;
-  }
-
-private:
-  /// Independent solution to the radial equation
-  IndependentSolution<StateVariable, ODESystem> solution_;
-};
-
-/*! \file InterfacesImpl.hpp
- *  \class Zeta
- *  \brief 1st solution to the radial second order ODE, with r^l behaviour in the
- * origin
- *  \author Roberto Di Remigio
- *  \date 2015
- *  \tparam StateVariable type of the state variable used in the ODE solver
- *  \tparam ODESystem system of 1st order ODEs replacing the 2nd order ODE
- */
-template <typename StateVariable, typename ODESystem> class Zeta __final {
-public:
-  Zeta() : L_(0), y_0_(0.0), y_infinity_(0.0) {}
-  Zeta(int l,
-       double y0,
-       double yinf,
-       const ProfileEvaluator & eval,
-       const IntegratorParameters & parms)
-      : L_(l), y_0_(y0), y_infinity_(yinf) {
-    compute(eval, parms);
-  }
-  ~Zeta() {}
-  pcm::tuple<double, double> operator()(double point) const {
-    return pcm::make_tuple(function_impl(point), derivative_impl(point));
-  }
-  friend std::ostream & operator<<(std::ostream & os, Zeta & obj) {
     for (size_t i = 0; i < obj.function_[0].size(); ++i) {
       os << obj.function_[0][i] << "    " << obj.function_[1][i] << "    "
          << obj.function_[2][i] << std::endl;
@@ -185,205 +158,113 @@ public:
   }
 
 private:
-  typedef pcm::array<StateVariable, 3> RadialSolution;
-  /// Angular momentum of the solution
+  typedef typename ODE::StateType StateType;
+  /*! Angular momentum of the solution */
   int L_;
-  /// Lower bound of the integration interval
-  double y_0_;
-  /// Upper bound of the integration interval
-  double y_infinity_;
-  /// The actual data: grid, function value and first derivative values
-  RadialSolution function_;
+  /*! Lower bound of the integration interval */
+  double y_min_;
+  /*! Upper bound of the integration interval */
+  double y_max_;
+  /*! The sign of the integration interval */
+  double y_sign_;
+  /*! The actual data: grid, function value and first derivative values */
+  pcm::array<std::vector<double>, 3> function_;
   /*! Reports progress of differential equation integrator */
-  void push_back(const StateVariable & x, double y) {
+  void push_back(const StateType & x, double y) {
     function_[0].push_back(y);
     function_[1].push_back(x[0]);
     function_[2].push_back(x[1]);
   }
-  /*! \brief Calculates 1st radial solution, i.e. the one with r^l behavior
+  /*! \brief Calculates radial solution
    *  \param[in] eval   dielectric profile evaluator function object
    *  \param[in] parms parameters for the integrator
+   *
+   *  This function discriminates between the first, i.e. the one with r^l
+   *  behavior, and the second radial solution, i.e. the one with r^(-l-1)
+   *  behavior, based on the sign of the integration interval y_sign_.
    */
   void compute(const ProfileEvaluator & eval, const IntegratorParameters & parms) {
     namespace odeint = boost::numeric::odeint;
-    odeint::runge_kutta4<StateVariable> stepper;
+    odeint::runge_kutta4<StateType> stepper;
 
-    ODESystem system(eval, L_);
+    ODE system(eval, L_);
     // Holds the initial conditions
-    StateVariable init_zeta(2);
+    StateType init;
     // Set initial conditions
-    init_zeta[0] = L_ * y_0_;
-    init_zeta[1] = L_;
+    if (y_sign_ > 0.0) { // zeta-type solution
+      init[0] = y_sign_ * L_ * y_min_;
+      init[1] = y_sign_ * L_;
+    } else { // omega-type solution
+      init[0] = y_sign_ * (L_ + 1) * y_min_;
+      init[1] = y_sign_ * (L_ + 1);
+    }
     odeint::integrate_const(
         stepper,
         system,
-        init_zeta,
-        y_0_,
-        y_infinity_,
-        parms.observer_step_,
-        pcm::bind(
-            &Zeta<StateVariable, ODESystem>::push_back, this, pcm::_1, pcm::_2));
-  }
-  /*! \brief Returns value of function at given point
-   *  \param[in] point evaluation point
-   *
-   *  We first check if point is below y_0_, if yes we use
-   *  the asymptotic form L*y in point.
-   */
-  double function_impl(double point) const {
-    double zeta = 0.0;
-    if (point <= y_0_) {
-      zeta = L_ * point;
-    } else {
-      zeta = utils::splineInterpolation(point, function_[0], function_[1]);
-    }
-    return zeta;
-  }
-  /*! \brief Returns value of 1st derivative of function at given point
-   *  \param[in] point evaluation point
-   *
-   *  Below y_0_, the as asymptotic form L is used. Otherwise we interpolate.
-   */
-  double derivative_impl(double point) const {
-    double zeta = 0.0;
-    if (point <= y_0_) {
-      zeta = L_;
-    } else {
-      zeta = utils::splineInterpolation(point, function_[0], function_[2]);
-    }
-    return zeta;
-  }
-};
-
-/*! \file InterfacesImpl.hpp
- *  \class Omega
- *  \brief 2nd solution to the radial second order ODE, with r^(-l-1) behaviour at
- * infinity
- *  \author Roberto Di Remigio
- *  \date 2015
- *  \tparam StateVariable type of the state variable used in the ODE solver
- *  \tparam ODESystem system of 1st order ODEs replacing the 2nd order ODE
- */
-template <typename StateVariable, typename ODESystem> class Omega __final {
-public:
-  Omega() : L_(0), y_0_(0.0), y_infinity_(0.0) {}
-  Omega(int l,
-        double y0,
-        double yinf,
-        const ProfileEvaluator & eval,
-        const IntegratorParameters & parms)
-      : L_(l), y_0_(y0), y_infinity_(yinf) {
-    compute(eval, parms);
-  }
-  ~Omega() {}
-  pcm::tuple<double, double> operator()(double point) const {
-    return pcm::make_tuple(function_impl(point), derivative_impl(point));
-  }
-  friend std::ostream & operator<<(std::ostream & os, Omega & obj) {
-    for (size_t i = 0; i < obj.function_[0].size(); ++i) {
-      os << obj.function_[0][i] << "    " << obj.function_[1][i] << "    "
-         << obj.function_[2][i] << std::endl;
-    }
-    return os;
-  }
-
-private:
-  typedef pcm::array<StateVariable, 3> RadialSolution;
-  /// Angular momentum of the solution
-  int L_;
-  /// Lower bound of the integration interval
-  double y_0_;
-  /// Upper bound of the integration interval
-  double y_infinity_;
-  /// The actual data: grid, function value and first derivative values
-  RadialSolution function_;
-  /*! Reports progress of differential equation integrator */
-  void push_back(const StateVariable & x, double y) {
-    function_[0].push_back(y);
-    function_[1].push_back(x[0]);
-    function_[2].push_back(x[1]);
-  }
-  /*! \brief calculates 2nd radial solution, i.e. the one with r^(-l-1) behavior
-   *  \param[in] eval   dielectric profile evaluator function object
-   *  \param[in] parms parameters for the integrator
-   */
-  void compute(const ProfileEvaluator & eval, const IntegratorParameters & parms) {
-    namespace odeint = boost::numeric::odeint;
-    odeint::runge_kutta4<StateVariable> stepper;
-
-    ODESystem system(eval, L_);
-    // Holds the initial conditions
-    StateVariable init_omega(2);
-    // Set initial conditions
-    init_omega[0] = -(L_ + 1) * y_infinity_;
-    init_omega[1] = -(L_ + 1);
-    // Notice that we integrate BACKWARDS, so we pass -step to integrate_adaptive
-    odeint::integrate_const(
-        stepper,
-        system,
-        init_omega,
-        y_infinity_,
-        y_0_,
-        -parms.observer_step_,
-        pcm::bind(
-            &Omega<StateVariable, ODESystem>::push_back, this, pcm::_1, pcm::_2));
-// Reverse order of StateVariable-s in RadialSolution
-// this ensures that they are in ascending order, as later expected by
-// function_impl and derivative_impl
+        init,
+        y_min_,
+        y_max_,
+        y_sign_ * parms.observer_step_,
+        pcm::bind(&RadialFunction<ODE>::push_back, this, pcm::_1, pcm::_2));
+    // clang-format off
+    // Reverse order of function_ if omega-type solution was computed
+    // this ensures that they are in ascending order, as later expected by
+    // function_impl and derivative_impl
+    if (y_sign_ < 0.0) {
 #ifdef HAS_CXX11
-    for (auto & comp : function_) {
+      for (auto & comp : function_) {
 #else  /* HAS_CXX11 */
-    BOOST_FOREACH (StateVariable & comp, function_) {
+      BOOST_FOREACH (std::vector<double> & comp, function_) {
 #endif /* HAS_CXX11 */
-      std::reverse(comp.begin(), comp.end());
+        std::reverse(comp.begin(), comp.end());
+      }
+      }
     }
-  }
+  // clang-format on
   /*! \brief Returns value of function at given point
    *  \param[in] point evaluation point
    *
-   * We first check if point is above r_infinity_, if yes we use
-   * the asymptotic form -(L+1)*y in point.
+   *  We first check if point is below y_min_, if yes we use
+   *  the asymptotic form.
    */
   double function_impl(double point) const {
-    double omega = 0.0;
-    if (point >= y_infinity_) {
-      omega = -(L_ + 1) * point;
+    double val = 0.0;
+    if (point <= y_min_ && y_sign_ > 0.0) { // Asymptotic zeta-type solution
+      val = y_sign_ * L_ * point;
+    } else if (point >= y_min_ && y_sign_ < 0.0) { // Asymptotic omega-type solution
+      val = y_sign_ * (L_ + 1) * point;
     } else {
-      omega = utils::splineInterpolation(point, function_[0], function_[1]);
+      val = utils::splineInterpolation(point, function_[0], function_[1]);
     }
-    return omega;
+    return val;
   }
   /*! \brief Returns value of 1st derivative of function at given point
    *  \param[in] point evaluation point
    *
-   * We first check if point is above r_infinity_, if yes we use
-   * the asymptotic form -(L+1) in point.
+   *  Below y_min_, the asymptotic form is used. Otherwise we interpolate.
    */
   double derivative_impl(double point) const {
-    double omega = 0.0;
-    if (point >= y_infinity_) {
-      omega = -(L_ + 1);
+    double val = 0.0;
+    if (point <= y_min_ && y_sign_ > 0.0) { // Asymptotic zeta-type solution
+      val = y_sign_ * L_;
+    } else if (point >= y_min_ && y_sign_ < 0.0) { // Asymptotic omega-type solution
+      val = y_sign_ * (L_ + 1);
     } else {
-      omega = utils::splineInterpolation(point, function_[0], function_[2]);
+      val = utils::splineInterpolation(point, function_[0], function_[2]);
     }
-    return omega;
+    return val;
   }
 };
 
 /*! \brief Write contents of a RadialFunction to file
- *  \param[in] f RadialSolution whose contents have to be printed
+ *  \param[in] f solution to the radial 2nd order ODE to print
  *  \param[in] fname name of the file
  *  \author Roberto Di Remigio
- *  \date 2015
- *  \tparam StateVariable type of the state variable used in the ODE solver
- *  \tparam ODESystem system of 1st order ODEs replacing the 2nd order ODE
- *  \tparam IndependentSolution encodes which type of radial solution
+ *  \date 2018
+ *  \tparam ODE system of 1st order ODEs replacing the 2nd order ODE
  */
-template <typename StateVariable,
-          typename ODESystem,
-          template <typename, typename> class IndependentSolution>
-void writeToFile(RadialFunction<StateVariable, ODESystem, IndependentSolution> & f,
-                 const std::string & fname) {
+template <typename ODE>
+void writeToFile(RadialFunction<ODE> & f, const std::string & fname) {
   std::ofstream fout;
   fout.open(fname.c_str());
   fout << f << std::endl;